+++ src/headless-navigation.ts
@@ -1,5 +1,6 @@
 import HeadlessUi from "./headless-ui"
 import { keycodeEquals } from "./utils"
+import HeadlessButton from "./headless-button";
 /**
  * Class for navigation
  * @class
@@ -13,7 +14,7 @@ import { keycodeEquals } from "./utils"
  * @since 0.1.4
  * */
 class HeadlessNavigation extends HeadlessUi {
-  protected declare readonly button: HTMLButtonElement | null
+  protected declare readonly button: HTMLButtonElement | HeadlessButton | null
   protected declare readonly mainContainer: HTMLElement | null
 
   constructor() {
@@ -28,10 +29,16 @@ class HeadlessNavigation extends HeadlessUi {
    * */
   protected checkRequirements() {
     if (!this.button) {
-      throw new Error(`A button element with attribute "aria-controls" and "aria-expanded" is required`)
+      console.log(
+        `c% A button element with attribute "aria-controls" and "aria-expanded" is required`,
+        `color:red;background-color:pink;font-size:18px; padding: 3px; border-radius: 3px; border: red 1px solid;`
+      )
     }
     if (!this.mainContainer) {
-      throw new Error(`A div element with ID equals button "aria-controls" is required`)
+      console.log(
+        `c% A div element with ID equals button "aria-controls" is required`,
+        `color:red;background-color:pink;font-size:18px; padding: 3px; border-radius: 3px; border: red 1px solid;`
+      )
     }
   }